+++ Release Notes.md
@@ -35,6 +35,7 @@ SOFTWARE.
 * arm_chip_stm32: added wrapper for PWM tim/channels
 * arm_chip_stm32: refactored wrappers for TIM & GPIO names to what's mentioned here
 * arm_chip_xx: Families divided into files with respective fam name
+* arm_errors: added to sarmfsw
 * refactoring names of platform header files
 
 ## v1.0